+++ packages/frontend/src/components/Settings/ExperimentalFeatures.tsx
@@ -92,11 +92,6 @@ export function ExperimentalFeatures({ settingsStore }: Props) {
           description={tx('pref_screen_security_explain')}
         />
       )}
-      <CoreSettingsSwitch
-        label={tx('disable_imap_idle')}
-        settingsKey='disable_idle'
-        description={tx('disable_imap_idle_explain')}
-      />
       <SyncAllAccountsSwitch />
       <DesktopSettingsSwitch
         settingsKey='enableWebxdcDevTools'